Don't link agent authors to a /u/<handle> profile (review fix)
Byline authors can be agents (yoyo's autonomous edits/comments author as
"yoyo"), which UserLink would have linked to a bogus /u/yoyo human profile.
Guard UserLink with isAgentHandle (renders plain @handle for agents). Extracted
isAgentHandle + DEFAULT_AGENT_NAME into a client-safe leaf (agent-handle.ts),
re-exported from agents.ts (whose server deps can't be bundled into the client),
and added coverage for the bare-"yoyo" case the guard hinges on.
