+++ src/components/UserLink.tsx
@@ -1,11 +1,12 @@
 import Link from "next/link";
 import { profileHref } from "@/lib/links";
+import { isAgentHandle } from "@/lib/agent-handle";
 
 /**
- * A human handle rendered as `@handle` linking to its profile (`/u/<handle>`).
- * Use for plain-text byline/author sites that don't use the {@link Mark} chip
- * (e.g. comments, revision history). Agents are NOT users — don't use this for
- * agent ids.
+ * A handle rendered as `@handle`. Human handles link to their profile
+ * (`/u/<handle>`); agent authors (e.g. yoyo's autonomous edits/comments) render
+ * as plain text — they have no `/u/<handle>` profile. Use for plain-text
+ * byline/author sites that don't use the {@link Mark} chip (comments, revisions).
  */
 export function UserLink({
   handle,
@@ -14,6 +15,9 @@ export function UserLink({
   handle: string;
   className?: string;
 }) {
+  if (isAgentHandle(handle)) {
+    return <span className={className}>@{handle}</span>;
+  }
   return (
     <Link href={profileHref(handle)} className={className}>
       @{handle}
